Move save and reset buttons to main icon bar on channel programming page
These icons pretty much always appeared below the fold and required
scroll, which started to feel kind of awkward.
