+++ web/src/pages/channels/ChannelProgrammingPage.tsx
@@ -1,63 +1,17 @@
-import { Save } from '@mui/icons-material';
-import {
-  Box,
-  Button,
-  CircularProgress,
-  Paper,
-  Snackbar,
-  Typography,
-} from '@mui/material';
-import { useTheme } from '@mui/material/styles';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { UpdateChannelProgrammingRequest } from '@tunarr/types/api';
-import { ZodiosError } from '@zodios/core';
-import { chain, findIndex, first, isUndefined, map } from 'lodash-es';
-import { useEffect, useState } from 'react';
+import { CircularProgress, Paper, Typography } from '@mui/material';
+import { isUndefined } from 'lodash-es';
+import { useEffect } from 'react';
 import Breadcrumbs from '../../components/Breadcrumbs.tsx';
 import { ChannelProgrammingConfig } from '../../components/channel_config/ChannelProgrammingConfig.tsx';
 import UnsavedNavigationAlert from '../../components/settings/UnsavedNavigationAlert.tsx';
-import { channelProgramUniqueId } from '../../helpers/util.ts';
 import { usePreloadedChannelEdit } from '../../hooks/usePreloadedChannel.ts';
-import { useTunarrApi } from '../../hooks/useTunarrApi.ts';
-import { useUpdateChannel } from '../../hooks/useUpdateChannel.ts';
-import {
-  resetCurrentLineup,
-  resetLineup,
-} from '../../store/channelEditor/actions.ts';
+import { resetLineup } from '../../store/channelEditor/actions.ts';
 import useStore from '../../store/index.ts';
 
-type MutateArgs = {
-  channelId: string;
-  lineupRequest: UpdateChannelProgrammingRequest;
-};
-
-type SnackBar = {
-  display: boolean;
-  message: string;
-  color: string;
-};
-
 export default function ChannelProgrammingPage() {
-  const {
-    currentEntity: channel,
-    originalEntity: originalChannel,
-    programList: newLineup,
-  } = usePreloadedChannelEdit();
+  const { currentEntity: channel } = usePreloadedChannelEdit();
 
   const programsDirty = useStore((s) => s.channelEditor.dirty.programs);
-
-  const apiClient = useTunarrApi();
-  const queryClient = useQueryClient();
-  const theme = useTheme();
-  const [isSubmitting, setIsSubmitting] = useState<boolean>(false);
-
-  const [snackStatus, setSnackStatus] = useState<SnackBar>({
-    display: false,
-    color: '',
-    message: '',
-  });
-
-  const updateChannelMutation = useUpdateChannel(/*isNewChannel=*/ false);
 
   // Force this page to load at the top
   // Fixes issue where some browsers maintain previous page scroll position
@@ -66,104 +20,12 @@ export default function ChannelProgrammingPage() {
     window.scrollTo(0, 0);
   }, []);
 
-  const updateLineupMutation = useMutation({
-    mutationFn: ({ channelId, lineupRequest }: MutateArgs) => {
-      return apiClient.post('/api/channels/:id/programming', lineupRequest, {
-        params: { id: channelId },
-      });
-    },
-    onSuccess: async (data, { channelId: channelNumber }) => {
-      resetCurrentLineup(data.lineup, data.programs);
-      setIsSubmitting(false);
-      setSnackStatus({
-        display: true,
-        message: 'Programs Saved!',
-        color: theme.palette.success.main,
-      });
-      await queryClient.invalidateQueries({
-        queryKey: ['channels', channelNumber],
-      });
-    },
-    onError: (error) => {
-      setIsSubmitting(false);
-      setSnackStatus({
-        display: true,
-        message: error.message,
-        color: theme.palette.error.main,
-      });
-      if (error instanceof ZodiosError) {
-        console.error(error.data);
-        console.error(error, error.cause);
-      }
-    },
-  });
-
-  const handleSnackClose = () => {
-    setSnackStatus({ display: false, message: '', color: '' });
-  };
-
-  const onSave = () => {
-    setIsSubmitting(true);
-    if (
-      !isUndefined(channel) &&
-      !isUndefined(originalChannel) &&
-      channel.startTime !== originalChannel.startTime
-    ) {
-      updateChannelMutation.mutate({
-        ...channel,
-        // This is a little wonky...
-        transcoding: {
-          targetResolution: channel.transcoding?.targetResolution ?? 'global',
-          videoBitrate: channel.transcoding?.videoBitrate ?? 'global',
-          videoBufferSize: channel.transcoding?.videoBufferSize ?? 'global',
-        },
-      });
-    }
-
-    // Group programs by their unique ID. This will disregard their durations,
-    // but we will keep the durations when creating the minimal lineup below
-    const uniquePrograms = chain(newLineup)
-      .groupBy((lineupItem) => channelProgramUniqueId(lineupItem))
-      .values()
-      .map(first)
-      .compact()
-      .value();
-
-    // Create the in-order lineup which is a lookup array - we have the index
-    // to the actual program (in the unique programs list) and then the
-    // duration of the lineup item.
-    const lineup = map(newLineup, (lineupItem) => {
-      const index = findIndex(
-        uniquePrograms,
-        (uniq) =>
-          channelProgramUniqueId(lineupItem) === channelProgramUniqueId(uniq),
-      );
-      return { duration: lineupItem.duration, index };
-    });
-
-    updateLineupMutation
-      .mutateAsync({
-        channelId: channel!.id,
-        lineupRequest: { type: 'manual', lineup, programs: uniquePrograms },
-      })
-      .then(console.log)
-      .catch(console.error);
-  };
-
   return isUndefined(channel) ? (
     <div>
       <CircularProgress />
     </div>
   ) : (
     <div>
-      <Snackbar
-        open={snackStatus.display}
-        autoHideDuration={6000}
-        anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}
-        onClose={handleSnackClose}
-        message={snackStatus.message}
-        sx={{ backgroundColor: snackStatus.color }}
-      />
       <Breadcrumbs />
       <Typography variant="h4" sx={{ mb: 2 }}>
         {channel.name}
@@ -174,36 +36,6 @@ export default function ChannelProgrammingPage() {
           isDirty={programsDirty}
           onProceed={() => resetLineup()}
         />
-        <Box
-          sx={{ display: 'flex', justifyContent: 'end', pt: 1, columnGap: 1 }}
-        >
-          {programsDirty && (
-            <Button
-              variant="contained"
-              onClick={() => resetLineup()}
-              disabled={!programsDirty}
-            >
-              Reset Changes
-            </Button>
-          )}
-          <Button
-            variant="contained"
-            onClick={() => onSave()}
-            disabled={!programsDirty || isSubmitting}
-            startIcon={
-              isSubmitting ? (
-                <CircularProgress
-                  size="20px"
-                  sx={{ mx: 1, color: 'inherit' }}
-                />
-              ) : (
-                <Save />
-              )
-            }
-          >
-            Save
-          </Button>
-        </Box>
       </Paper>
     </div>
   );
